Create README.md
ViewModel: Knockout's ViewModel is similar to the Octopus. It separates the Model and the View

Declarative Bindings: Bindings allow you to connect the View and Model in a direct and simple way.

Automatic UI Refresh: Knockout's will update the View when the Model changes. And with the right declarative bindings, Knockout can update the Model when elements in the View change (such as input elements, checkboxes, etc).

Dependency Tracking: Knockout allows you to create a relationship between parts of the Model, and will automatically update Model data that depends on other Model data when that other Model data changes.